@@ -11,36 +11,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@angular/cdk": "^5.0.0",
-    "@angular/material": "^5.0.0",
+    "@angular/animations": "^6.1.2",
+    "@angular/cdk": "^6.4.5",
+    "@angular/common": "^6.1.2",
+    "@angular/compiler": "^6.1.2",
+    "@angular/core": "^6.1.2",
+    "@angular/forms": "^6.1.2",
+    "@angular/http": "^6.1.2",
+    "@angular/material": "^6.4.5",
+    "@angular/platform-browser": "^6.1.2",
+    "@angular/platform-browser-dynamic": "^6.1.2",
+    "@angular/router": "^6.1.2",
     "angular2-uuid": "^1.1.1",
-    "angularfire2": "5.0.0-rc.2",
+    "angularfire2": "5.0.0-rc.11",
     "bootstrap": "^4.1.1",
     "core-js": "^2.4.1",
-    "firebase": "^4.5.0",
+    "firebase": "5.3.1",
     "font-awesome": "^4.7.0",
-    "rxjs": "^5.5.2",
-    "typescript": "~2.4.2",
+    "rxjs": "^6.2.2",
+    "typescript": "2.9.1",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.1",
-    "@angular/cli": "^6.0.1",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular/cli": "6.1.3",
+    "@angular/compiler-cli": "^6.1.2",
+    "@angular/language-service": "^6.1.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
+    "codelyzer": "4.4.4",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -51,7 +51,6 @@
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "tslint": "~5.7.0"
   }
 }
